binanceapp/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
dateparser==1.1.0
fonttools==4.28.5
frozenlist==1.2.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==5.2.0
numpy==1.22.1
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.6
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.27.1
six==1.16.0
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
